:recycle: 자잘한 코드 정리
