+++ lib/presentation/my_page/update_profile/update_profile_info_view.dart
@@ -16,41 +16,20 @@ import 'package:frontend_mobile/presentation/global/user/user_view_model.dart';
 import 'package:frontend_mobile/presentation/my_page/update_profile/update_profile_info_view_model.dart';
 import 'package:go_router/go_router.dart';
 
-class UpdateProfileInfoView extends ConsumerStatefulWidget {
+part 'local_widget/profile_image_and_nickname.dart';
+part 'local_widget/preferences.dart';
+part 'local_widget/save_button.dart';
+
+class UpdateProfileInfoView extends ConsumerWidget {
   const UpdateProfileInfoView({super.key});
 
   @override
-  ConsumerState<UpdateProfileInfoView> createState() =>
-      _UpdateProfileInfoViewState();
-}
-
-class _UpdateProfileInfoViewState extends ConsumerState<UpdateProfileInfoView> {
-  final TextEditingController _nicknameController = TextEditingController();
-  final String _successText = '사용 가능한 닉네임입니다';
-  final String _errorText = '사용 불가능한 닉네임입니다';
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) async {
-      final UserState userState = ref.read(userViewModelProvider);
-      _nicknameController.text = userState.data.nickname;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final UserState userState = ref.watch(userViewModelProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
     final UpdateProfileInfoState state = ref.watch(
       updateProfileInfoViewModelProvider,
     );
-    final PreferenceState preferences = ref.watch(preferenceViewModelProvider);
-    final UpdateProfileInfoViewModel viewmodel = ref.read(
-      updateProfileInfoViewModelProvider.notifier,
-    );
 
     final ColorScheme colorScheme = Theme.of(context).colorScheme;
-    final TextTheme textTheme = Theme.of(context).textTheme;
 
     ref.listen(
       updateProfileInfoViewModelProvider.select(
@@ -82,214 +61,17 @@ class _UpdateProfileInfoViewState extends ConsumerState<UpdateProfileInfoView> {
               Expanded(
                 child: ListView(
                   children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.only(
-                        left: 16,
-                        top: 25,
-                        right: 16,
-                        bottom: 50,
-                      ),
-                      child: Column(
-                        children: <Widget>[
-                          Align(
-                            child:
-                                userState.isMale
-                                    ? CustomProfilePhotoSetting.boy(
-                                      onCameraTap: (Uint8List image) {
-                                        viewmodel.updateProfileImage(
-                                          imageUint8List: image,
-                                        );
-                                      },
-                                      imageUrl: state.oldImageUrl,
-                                    )
-                                    : CustomProfilePhotoSetting.girl(
-                                      onCameraTap: (Uint8List image) {
-                                        viewmodel.updateProfileImage(
-                                          imageUint8List: image,
-                                        );
-                                      },
-                                      imageUrl: state.oldImageUrl,
-                                    ),
-                          ),
-                          const SizedBox(height: 26),
-                          Row(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Expanded(
-                                child: CustomInputBox(
-                                  controller: _nicknameController,
-                                  hintText: '닉네임',
-                                  closeControll: true,
-                                  success:
-                                      state.nicknameAvailabilityData.available,
-                                  successText: _successText,
-                                  error:
-                                      state.postNicknameStatus.isFailure ||
-                                      (state.postNicknameStatus.isSuccess &&
-                                          !state
-                                              .nicknameAvailabilityData
-                                              .available),
-                                  errorText: _errorText,
-                                  onCloseButtonTap: () {
-                                    viewmodel.updateNickName(nickname: '');
-                                  },
-                                  onChanged: (String val) {
-                                    viewmodel.updateNickName(nickname: val);
-                                  },
-                                ),
-                              ),
-                              const SizedBox(width: 10),
-                              CustomFillButton.medium(
-                                label: '중복확인',
-                                backgroundColor: CustomFillButtonColor.olive,
-                                disabled: state.nickName.isEmpty,
-                                width: 110,
-                                onPressed: () {
-                                  viewmodel.postNickname();
-                                },
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
+                    const _ProfileImageAndNickname(),
                     Divider(height: 1, color: colorScheme.outline),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 16,
-                        vertical: 15,
-                      ),
-                      child: Row(
-                        children: <Widget>[
-                          Expanded(
-                            child: Text(
-                              '내 디저트 취향',
-                              style: textTheme.titleSmall?.copyWith(
-                                color: ScaleColorConfig.primary20,
-                              ),
-                            ),
-                          ),
-                          RichText(
-                            text: TextSpan(
-                              children: <InlineSpan>[
-                                TextSpan(
-                                  text: state.userPreferences.length.toString(),
-                                  style: textTheme.labelLarge?.copyWith(
-                                    color: ScaleColorConfig.success50,
-                                  ),
-                                ),
-                                TextSpan(
-                                  text: '개 선택',
-                                  style: textTheme.labelLarge?.copyWith(
-                                    color: ScaleColorConfig.neutral30,
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    Divider(height: 1, color: colorScheme.outline),
-                    Padding(
-                      padding: const EdgeInsets.all(16),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: <Widget>[
-                          _PreferenceFilters(
-                            title: '식단&건강',
-                            preferences: preferences.data.sublist(0, 5),
-                          ),
-                          const SizedBox(height: 24),
-                          _PreferenceFilters(
-                            title: '트렌드&감성',
-                            preferences: preferences.data.sublist(5, 8),
-                          ),
-                          const SizedBox(height: 24),
-                          _PreferenceFilters(
-                            title: '경험&독창성',
-                            preferences: preferences.data.sublist(8),
-                          ),
-                        ],
-                      ),
-                    ),
+                    const _Preferences(),
                   ],
                 ),
               ),
-              Container(
-                decoration: BoxDecoration(
-                  border: Border(top: BorderSide(color: colorScheme.outline)),
-                ),
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 16,
-                  vertical: 20,
-                ),
-                child: CustomFillButton.medium(
-                  label: '저장',
-                  disabled:
-                      !state.postNicknameStatus.isSuccess ||
-                      !state.nicknameAvailabilityData.available,
-                  onPressed: () {
-                    viewmodel.updateProfileInfo();
-                  },
-                ),
-              ),
+              const _SaveButton(),
             ],
           ),
         ),
       ),
-    );
-  }
-
-  @override
-  void dispose() {
-    _nicknameController.dispose();
-    super.dispose();
-  }
-}
-
-// 유저 취향 필터 리스트
-class _PreferenceFilters extends ConsumerWidget {
-  const _PreferenceFilters({required this.preferences, required this.title});
-  final String title;
-  final List<PreferenceModel> preferences;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final TextTheme textTheme = Theme.of(context).textTheme;
-    final UpdateProfileInfoState state = ref.watch(
-      updateProfileInfoViewModelProvider,
-    );
-    final UpdateProfileInfoViewModel viewmodel = ref.read(
-      updateProfileInfoViewModelProvider.notifier,
-    );
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text(
-          title,
-          style: textTheme.labelLarge?.copyWith(
-            color: ScaleColorConfig.neutral40,
-          ),
-        ),
-        const SizedBox(height: 12),
-        Wrap(
-          runSpacing: 6,
-          spacing: 6,
-          children: <Widget>[
-            ...preferences.map(
-              (PreferenceModel e) => CustomSuggestiveChip(
-                label: e.preferenceName,
-                onPressed: () {
-                  viewmodel.updateUserPreferences(preferenceTagId: e.id);
-                },
-                isSelected: state.userPreferences.contains(e.id),
-              ),
-            ),
-          ],
-        ),
-      ],
     );
   }
 }
